+++ core-ui/src/components/Clusters/components/KubeconfigUpload/KubeconfigTextArea/KubeconfigTextArea.scss
@@ -6,4 +6,8 @@
         resize: none;
         width: 100%;
     }
+
+    button {
+        width: 100%;
+    }
 }
