+++ core-ui/src/components/Clusters/components/KubeconfigUpload/KubeconfigUpload.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { MessageStrip, Tab, TabGroup } from 'fundamental-react';
+import { MessageStrip } from 'fundamental-react';
 import { KubeconfigFileUpload } from './KubeconfigFileUpload';
 import { KubeconfigTextArea } from './KubeconfigTextArea/KubeconfigTextArea';
 import jsyaml from 'js-yaml';
@@ -8,16 +8,10 @@ export function KubeconfigUpload({
   handleKubeconfigAdded,
   kubeconfigFromParams,
 }) {
-  const [tabIndex, setTabIndex] = React.useState(0);
   const [showParseError, setShowParseError] = React.useState(false);
   const [kubeconfigs, setKubeconfigs] = React.useState({
     text: jsyaml.dump(kubeconfigFromParams),
   });
-
-  React.useEffect(() => {
-    // select second tab, the one with text field
-    if (!!kubeconfigFromParams) setTabIndex(1);
-  }, [kubeconfigFromParams]);
 
   const parseKubeconfig = text => {
     try {
@@ -41,27 +35,14 @@ export function KubeconfigUpload({
 
   return (
     <>
-      <TabGroup
-        selectedIndex={tabIndex}
-        onTabClick={(_, index) => {
-          setShowParseError(false);
-          handleKubeconfigAdded(kubeconfigs[index === 0 ? 'upload' : 'text']);
-        }}
-        className="fd-margin-bottom--sm"
-      >
-        <Tab title="Upload">
-          <KubeconfigFileUpload
-            onKubeconfigTextAdded={onKubeconfigTextAdded('upload')}
-          />
-        </Tab>
-        <Tab title="Paste">
-          <KubeconfigTextArea
-            onKubeconfigTextAdded={onKubeconfigTextAdded('text')}
-            kubeconfigFromParams={kubeconfigFromParams}
-          />
-        </Tab>
-      </TabGroup>
-
+      <KubeconfigFileUpload
+        onKubeconfigTextAdded={onKubeconfigTextAdded('upload')}
+      />
+      <p>or</p>
+      <KubeconfigTextArea
+        onKubeconfigTextAdded={onKubeconfigTextAdded('text')}
+        kubeconfigFromParams={kubeconfigFromParams}
+      />
       {showParseError && (
         <MessageStrip
           aria-label="invalid-kubeconfig"
