+++ tests/fixtures/kubeconfig.yaml
@@ -1,1 +1,19 @@
-wrong_kubeconfig
+apiVersion: v1
+kind: Config
+clusters:
+  - name: shoot--hasselhoff--kmain
+    cluster:
+      server: 'https://api.kmain.hasselhoff.shoot.canary.k8s-hana.ondemand.com'
+      certificate-authority-data: >-
+        LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUM5akNDQWQ2Z0F3SUJBZ0lRQXJsNFhFQlVENktCSVJLTVY2NTg0VEFOQmdrcWhraUc5dzBCQVFzRkFEQVYKTVJNd0VRWURWUVFERXdwcmRXSmxjbTVsZEdWek1CNFhEVEl4TURZd01qQTBNVEl4TVZvWERUTXhNRFl3TWpBMApNVEl4TVZvd0ZURVRNQkVHQTFVRUF4TUthM1ZpWlhKdVpYUmxjekNDQVNJd0RRWUpLb1pJaHZjTkFRRUJCUUFECmdnRVBBRENDQVFvQ2dnRUJBTll3bXppZkFlZWZ6MGwrQ0s0OFplaWc4VG9CM3cxYVgyR1hvN015SVVEcWUwdUwKNmhoNHlIOVhodFNtZy9qVHpyU1Bwb0gzVllWczljVGtXQ21XOFpLWEZyQ2RUQTZTUThEK0FvQVNFWlBmR2x6MgpPMjAwbmhzeHlHMGtsdFpFNXBpT2tWaXFoc3JLdm01WTFnR0N0STBkaVhKVWcxWjQvOEdMM0NVbkxMSU93aDUyCjhTWXdDWXZCQW9za09SaXZkMlByZ0luNUEwayt5ZWFLZXpURHhKajBRWGY1Z1QvamFDWE5UUDNVNFFHSVVFSVUKb1NlYk85OWN3MDNuWUNrNUQ0UFYvUEhWYXpIOUMwNFRnTFFCYTcycENZS0hXU1dHUW9TK2NGczllSnRlUlBLUQp6U0JreXBiUC9qWC9GWVhERzRGYmFqMGQ4ZG9GdlBHaHBwMTBlczhDQXdFQUFhTkNNRUF3RGdZRFZSMFBBUUgvCkJBUURBZ0dtTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRkNpRGE0UHNBUW0wdjhUeXE0WTQKbXpwazFPL3RNQTBHQ1NxR1NJYjNEUUVCQ3dVQUE0SUJBUURDNFljUDJBbDhXOHIvc0dnZ05mNEdQL3ZZSEVYcwpWR0pGSlZQSXR3ZC95RGQ3UXFkQ0s3VlF2aXEzVy9Sb3FubXI4eit5bmJKemFoajRZQjZoajRrTFBZeUdPbGozCnYwLzF2WXpNYk1icnlZM01KNkxjaXBiOE80MGpiVUd3OVlCT3FMcUo5N3NuNFJIT0RaaGRPSjdsNk8vRTgzS0IKRzBodkp5OWRsWkM1U01HeHZRa3ViUng2MjdJYmVpUlJyOXpyYW1WNW5qL2lHc0ozYmZiNk12d1c4RjhXWktIbwpQdXZ2SktwSWgySmw2KzBzTmc2d0FIZ1hJSlVyR2pUTlBweTBqdlh3TEQ1TGdvRm1DZ2NlSEpzM00yd01HODNMCmxvN0dSTlc0UDhadUNFSC9qdElhVkRDN2c1alpXVGJ3WkFoYVhJTkNVL3JSVUFxcVhMRXVpOU5RCi0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0K
+contexts:
+  - name: shoot--hasselhoff--kmain
+    context:
+      cluster: shoot--hasselhoff--kmain
+      user: shoot--hasselhoff--kmain-token
+current-context: shoot--hasselhoff--kmain
+users:
+  - name: shoot--hasselhoff--kmain-token
+    user:
+      token: >-
+        5X3lmYkbOxtkH3SGnAbGTfbTQVcTwTzRAmPBfO8kXbWYBYzmrwupzYjLox42CSnNpciMaMbO1ALlOL1LMxgzW5m6HhzWUdcrMpqYX6cXIjfj7InntLVqbRZqxNpA5y0y
